Investments

+++ GenesisVision.Core/Services/Interfaces/ISmartContractService.cs
@@ -9,7 +9,7 @@ namespace GenesisVision.Core.Services.Interfaces
 {
     public interface ISmartContractService
     {
-        OperationResult RegisterManager(string tokenName, string tokenSymbol, string managerId, string managerLogin, string brokerId, uint managementFee, uint successFee); // TODO uint8
+        OperationResult RegisterManager(string tokenName, string tokenSymbol, string managerId, string managerLogin, string brokerId, decimal managementFee, decimal successFee); // TODO uint8
 
         OperationResult SetInitialTokensHolder(string holderAddress);
     }
